Add initial data files and core processing scripts
Added raw and cleaned flight data CSVs, session data, and main processing scripts including parser and visualization modules. This sets up the initial project structure for data ingestion, cleaning, segmentation, and plotting.

The cleaner module need to be improved to allow the suppression of the "last" raw that block the parser to work correcty. So for the moment, after the cleaner, you need to erase some columns on the "cleaned" file to allow the parser and ploter module to work. But on the actual commit, the figures will be generated.

Co-Authored-By: yohji79 <42109336+yohji79@users.noreply.github.com>